Attempt to fix broken treeIndex without botchering performance on iOS

+++ viewer/packages/rendering/src/glsl/base/updateFragmentColor.glsl
@@ -23,7 +23,7 @@ vec3 packNormalToRgb( const in vec3 normal ) {
 }
 
 void updateFragmentColor(
-    int renderMode, vec4 color, float treeIndex, 
+    int renderMode, vec4 color, int treeIndex, 
     vec3 normal, float depth, sampler2D matCapTexture, 
     int geometryType) {
     if (renderMode == RenderTypeColor || renderMode == RenderTypeEffects) {
